requirements.txt
Flask==2.3.3
flask-jwt-extended==4.4.4
pymongo==4.4.0
python-dotenv==1.0.0
gunicorn==20.1.0
werkzeug==2.3.7


